Introduce an extra branch for unreleased Swing features.
Previously `with-latest-snapshot-versions` was supposed to contain
examples for unreleased features of all AssertJ modules. Since at least
Core and Swing modules are not released in parallel, it's hard to merge
that branch to master when either Core or Swing module would be
released. Either way the master would contain examples for unreleased
features for the other module.

Therefore it might be a good idea to use several branches for tracking
the examples for new features for the different modules. Currently
`with-latest-swing-snapshot` has been added for the Swing module.