Upgrade & enhance existing Neo4j examples to v2
Node label case now follows the conventional one (i.e. `MyLabel`
instead of `MY_LABEL`).

Transactions are not committed in the examples as there is nothing
to commit (data is only read from the graph). They are also not
closed explicitly anymore as this is managed by the
try-with-resources block anyway.

Some repository methods were relying on an internal Neo4j order,
that is now fixed with an explicit Cypher `ORDER BY` clause.

Guava usage has been removed in favor of Java 8 API (and
assertions).